# apps/inference/openapi.json
{
  "components": {
    "schemas": {
      "ActivationAllBatchRequest": {
        "description": "For a given batch of prompts, get the top activating features for a set of SAEs (eg gemmascope-res-65k), or specific SAEs in the set of SAEs (eg 0-gemmascope-res-65k, 5-gemmascope-res-65k). Also has other customization options.",
        "properties": {
          "ignoreBos": {
            "description": "Whether or not to include features whose highest activation value is the BOS token.",
            "title": "Ignorebos",
            "type": "boolean"
          },
          "model": {
            "description": "Name of the model to test activations on",
            "title": "Model",
            "type": "string"
          },
          "numResults": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 25,
            "description": "Optional. The number of top features to return.",
            "title": "Numresults"
          },
          "prompts": {
            "description": "Input text prompts to get activations for",
            "items": {
              "type": "string"
            },
            "title": "Prompts",
            "type": "array"
          },
          "selectedSources": {
            "description": "List of specific SAEs to get activations for (eg [\"0-gemmascope-res-65k\", \"5-gemmascope-res-65k\"]). If not specified, will get activations for all SAEs in the source set.",
            "items": {
              "type": "string"
            },
            "title": "Selectedsources",
            "type": "array"
          },
          "sortByTokenIndexes": {
            "description": "Sort the results by the sum of the activations at the specified token indexes.",
            "items": {
              "type": "integer"
            },
            "title": "Sortbytokenindexes",
            "type": "array"
          },
          "sourceSet": {
            "description": "The source set name of the SAEs (eg gemmascope-res-16k)",
            "title": "Sourceset",
            "type": "string"
          }
        },
        "required": [
          "prompts",
          "model",
          "sourceSet",
          "selectedSources",
          "sortByTokenIndexes",
          "ignoreBos"
        ],
        "title": "ActivationAllBatchRequest",
        "type": "object"
      },
      "ActivationAllBatchResponse": {
        "description": "Response for NPActivationAllBatchRequest. Contains the batch results of activations for each top feature and the tokenized prompts.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ActivationAllBatchResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "ActivationAllBatchResponse",
        "type": "object"
      },
      "ActivationAllBatchResult": {
        "description": "One prompt's result within a batched top-features request.",
        "properties": {
          "activations": {
            "items": {
              "$ref": "#/components/schemas/ActivationAllFeature"
            },
            "title": "Activations",
            "type": "array"
          },
          "counts": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Not currently supported and may be incorrect. This is the number of features that activated by layer, starting from layer 0 of this SAE. Need to be redesigned.",
            "title": "Counts"
          },
          "tokens": {
            "items": {
              "type": "string"
            },
            "title": "Tokens",
            "type": "array"
          }
        },
        "required": [
          "activations",
          "tokens"
        ],
        "title": "ActivationAllBatchResult",
        "type": "object"
      },
      "ActivationAllFeature": {
        "description": "One feature and its activation in an NPActivationAllResponse",
        "properties": {
          "dfaMaxValue": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dfamaxvalue"
          },
          "dfaTargetIndex": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dfatargetindex"
          },
          "dfaValues": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dfavalues"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "maxValue": {
            "title": "Maxvalue",
            "type": "number"
          },
          "maxValueIndex": {
            "title": "Maxvalueindex",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "sumValues": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sumvalues"
          },
          "values": {
            "items": {
              "type": "number"
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "source",
          "index",
          "values",
          "maxValue",
          "maxValueIndex"
        ],
        "title": "ActivationAllFeature",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "SimpleSecretAuth": {
        "in": "header",
        "name": "X-SECRET-KEY",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "email": "johnny@neuronpedia.org"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Neuronpedia - Inference Server",
    "version": "1.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/activation/all-batch": {
      "post": {
        "operationId": "activationAllBatchPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationAllBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationAllBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activation All Batch"
      }
    },
...
  },
  "security": [
    {
      "SimpleSecretAuth": []
    }
  ]
}